+++ Source/DesignSystem/atoms/AccordionList/AccordionList.stories.tsx
@@ -2,16 +2,15 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 import React from 'react';
-import { Typography } from '@mui/material';
 
-import { componentStories, AccordionList  } from '@dolittle/design-system';
-import { AccordionProps } from '../Accordion/Accordion';
+import { componentStories, AccordionProps, AccordionList, MaxWidthTextBlock } from '@dolittle/design-system';
 
 const DummyChildrenContent = () =>
-    <Typography>
+    <MaxWidthTextBlock>
         Lorem ipsum dolor sit amet consectetur adipisicing elit. Culpa illo similique consequuntur dolores natus ex
         sunt esse mollitia id atque. Atque molestiae cum magnam eligendi maxime id sapiente quaerat suscipit?
-    </Typography>;
+    </MaxWidthTextBlock>;
+
 const dummyAccordionList: AccordionProps[] = [
     {
         id: '1',
@@ -35,38 +34,38 @@ const { metadata, createStory } = componentStories(AccordionList);
 metadata.parameters = {
     docs: {
         description: {
-            component: `The AlertList component is a convenient helper component
-that allows for easy composition of Accordion components. It's main use-case is
-to enable single-expand mode, where only a single Accordion component is expanded
-at a time.`
-        }
-    }
+            component: `The AlertList component is a convenient helper component that allows for easy composition 
+            of Accordion components. It's main use-case is to enable single-expand mode, where only a single Accordion 
+            component is expanded at a time.`
+        },
+    },
 };
-metadata.argTypes={
-    singleExpandMode: { control: true }
+
+metadata.argTypes = {
+    initialId: { control: false },
 };
 
 metadata.args = {
     items: dummyAccordionList,
-    singleExpandMode: true
+    singleExpandMode: false,
 };
-
 
 export default metadata;
 
 export const Default = createStory();
 
-export const InitialExpanded = createStory();
+export const InitialExpanded = createStory({
+    items: dummyAccordionList,
+    initialId: '2',
+});
 InitialExpanded.parameters = {
     docs: {
-        description:{
-            story:`In this example, you can see the second accordion item is pre-expanded`,
-        }
-    }
+        description: {
+            story: `In this example, you can see the second accordion item is pre-expanded`,
+        },
+    },
 };
 
-InitialExpanded.args = {
-    items: dummyAccordionList,
+export const SingleExpandMode = createStory({
     singleExpandMode: true,
-    initialId: '2'
-};
+});