+++ Source/SelfService/Web/applications/microservice/microservices/noMicroservices.tsx
@@ -3,8 +3,7 @@
 
 import React from 'react';
 
-import { Typography } from '@mui/material';
-import { RocketLaunch } from '@mui/icons-material';
+import { Icon } from '@dolittle/design-system';
 
 import { NoEntityView } from '../../../components/noEntityView/noEntityView';
 
@@ -17,15 +16,9 @@ export const NoMicroservices = ({ onCreate }: NoMicroservicesProps) =>
         title='No microservices deployed yet...'
         createEntityProps={{
             createEntityText: 'Deploy new microservice',
-            createEntityIcon: <RocketLaunch />,
-            onCreateEntity: onCreate
+            createEntityIcon: <Icon icon='RocketLaunch' />,
+            onCreateEntity: onCreate,
         }}
-    >
-        <Typography variant='body1' mb={2}>
-            After you deploy your first microservice it will appear here.
-        </Typography>
-
-        <Typography variant='body1'>
-            To deploy a new microservice click on the &apos;deploy a microservice&apos; button or &apos;deploy new&apos; tab at the top.
-        </Typography>
-    </NoEntityView>;
+        description='After you deploy your first microservice it will appear here.'
+        subDescription={`To deploy a new microservice click on the 'deploy a microservice' button or 'deploy new' tab at the top.`}
+    />;