+++ Source/SelfService/Web/components/noEntityView/noEntityView.tsx
@@ -5,6 +5,8 @@ import React from 'react';
 
 import { Grid, Typography } from '@mui/material';
 
+import { MaxWidthTextBlock } from '@dolittle/design-system';
+
 import { CreateEntityBox, CreateEntityBoxProps } from './createEntityBox';
 
 /**
@@ -17,18 +19,21 @@ export type NoEntityViewProps = {
     title: string;
 
     /**
-     * The props for the {@link CreateEntityBox} component*
+     * The props for the {@link CreateEntityBox} component.
      */
     createEntityProps: CreateEntityBoxProps;
 
     /**
      * Children to display addition text or information at the bottom of the view.
+     *
      * Use this space for descriptive text that helps explain what it means to not have an entity to the user.
      */
     children?: React.ReactNode;
+    description?: string;
+    subDescription?: string;
 };
 
-export const NoEntityView = ({ title, createEntityProps, children }: NoEntityViewProps) =>
+export const NoEntityView = ({ title, createEntityProps, description, subDescription, children }: NoEntityViewProps) =>
     <Grid
         container
         spacing={0}
@@ -41,8 +46,8 @@ export const NoEntityView = ({ title, createEntityProps, children }: NoEntityVie
 
         <CreateEntityBox {...createEntityProps} />
 
+        <MaxWidthTextBlock sx={{ mb: 2 }}>{description}</MaxWidthTextBlock>
+        <MaxWidthTextBlock>{subDescription}</MaxWidthTextBlock>
+
         {children}
     </Grid>;
-
-
-